Angular: Prevent angular from loading when disabled
