+++ public/app/core/navigation/errorModels.ts
@@ -1,0 +1,22 @@
+import { NavModel, NavModelItem } from '@grafana/data';
+
+export function getExceptionNav(error: unknown): NavModel {
+  console.error(error);
+  return getWarningNav('Exception thrown', 'See console for details');
+}
+
+export function getNotFoundNav(): NavModel {
+  return getWarningNav('Page not found', '404 Error');
+}
+
+export function getWarningNav(text: string, subTitle?: string): NavModel {
+  const node: NavModelItem = {
+    text,
+    subTitle,
+    icon: 'exclamation-triangle',
+  };
+  return {
+    node: node,
+    main: node,
+  };
+}
